Align hrlib's CMake structure
