+++ include/patheditor/path.hpp
@@ -23,7 +23,7 @@
 #ifndef PATH_HPP
 #define PATH_HPP
 
-#include "hrlib/fwd/qtfwd.h"
+#include "hrlib/fwd/qtfwd.hpp"
 #include "patheditor/fwd/patheditorfwd.hpp"
 
 #include <QObject>
