+++ include/patheditor/pathitem.hpp
@@ -23,7 +23,7 @@
 #ifndef PATHITEM_HPP
 #define PATHITEM_HPP
 
-#include "hrlib/fwd/qtfwd.h"
+#include "hrlib/fwd/qtfwd.hpp"
 #include "patheditor/fwd/patheditorfwd.hpp"
 
 #include <memory>
