+++ include/patheditor/pointcontextmenu.hpp
@@ -23,7 +23,7 @@
 #ifndef POINTCONTEXTMENU_HPP
 #define POINTCONTEXTMENU_HPP
 
-#include "hrlib/fwd/qtfwd.h"
+#include "hrlib/fwd/qtfwd.hpp"
 #include "patheditor/fwd/patheditorfwd.hpp"
 
 #include <QObject>
