+++ src/hrlib/CMakeLists.txt
@@ -1,11 +1,8 @@
 project(hrlib)
 
-include_directories(${CMAKE_SOURCE_DIR}/include/hrlib/hrlibfwd)
-include_directories(${CMAKE_SOURCE_DIR}/include/hrlib/math)
-include_directories(${CMAKE_SOURCE_DIR}/include/hrlib/mixin)
-include_directories(${CMAKE_SOURCE_DIR}/include/hrlib/string)
+file(GLOB_RECURSE HDR ${CMAKE_SOURCE_DIR}/include/hrlib/*.hpp)
 
-set(hrlibSRC
+set(SRC
     exceptions.cpp
     math/brent.cpp
     math/spline.cpp
@@ -13,6 +10,6 @@ set(hrlibSRC
     version.cpp
 )
 
-add_library(hrlib ${FINFOIL_LIB_TYPE} ${hrlibSRC} ${hrlibHDR})
+add_library(hrlib ${FINFOIL_LIB_TYPE} ${SRC} ${HDR})
 
 target_link_libraries(hrlib Qt5::Core)
