+++ tests/unittests/patheditor-tests/quadrantrestrictortest.cpp
@@ -23,7 +23,7 @@
 #include "quadrantrestrictortest.h"
 #include "patheditor/quadrantrestrictor.hpp"
 #include "submodules/qtestrunner/qtestrunner.hpp"
-#include "hrlib/exceptions.h"
+#include "hrlib/exceptions.hpp"
 
 void QuadrantRestrictorTest::testException()
 {
